Create Euclidean Algorithm
Attention: Use Recursion

The euclidean algorithm is very old, it was created as a means to calculate the Greatest Common Divisor (GCD) of two numbers. This algorithm is more efficient than calculating the GCD by factoring. We know that integer division is done as follows:

A = D*Q + R

where A is the dividend, D is the divisor, Q is the quotient of the integer division of A by D and R is the "mod" or remainder of the integer division.

Let's say that we want to calculate GCD(36,24).

36=24*1+12

Notice that 12 is a divisor of 36 and 24.

Therefore the GCD(36,24) = GCD(24,12)

24=12*2 + 0

Since the remainder is 0, we know 12 divides 24

GCD(36,24)=GCD(24,12)=12

In general, what the euclidean algorithm does is:

1. Choose the two numbers to calculate the GCD(A,D)
2. Divide A by D, if the remainder is not zero divide D by the remainder until the remainder found is 0.
3. When you find a division with remainder 0, the "D" of that division will be the GCD of the original values of A and D.